requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.4.26
et_xmlfile==2.0.0
frozenlist==1.6.0
idna==3.10
iso8601==2.1.0
magic-filter==1.0.12
multidict==6.4.3
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tortoise-orm==0.25.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
xlrd==2.0.1
yarl==1.20.0
